requirements.txt
Django~=4.2.7
djangorestframework~=3.14.0
amqp==5.1.1
anyio==4.0.0
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
decorator==5.1.1
django-celery-beat==2.5.0
django-cors-headers==4.3.0
django-filter==23.3
django-rest-framework==0.1.0
django-timezone-field==6.0.1
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
drf-yasg==1.21.7
eventlet==0.33.3
executing==2.0.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
inflection==0.5.1
ipython==8.16.1
jedi==0.19.1
kombu==5.3.2
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pyTelegramBotAPI==4.14.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8